requirements.txt
python-dateutil<2.8.1,>=2.1
Click>=6.7
PyYAML>=5.1
cassandra-driver>=3.14.0
psutil>=5.4.7
ffwd>=0.0.2
apache-libcloud>=2.8.0
lockfile>=0.12.2
pycrypto>=2.6.1
retrying>=1.3.3
parallel-ssh==1.9.1
requests==2.22.0
dnspython>=2.0.0
